ENH: Add "elxPrepareImageTypeSupport.h", adjust elxSupportedImageTypes.h
- Moved the definitions of `ElastixTypedef<VIndex>` and `elxSupportedImageTypeMacro` from "elxMacro.h" to "elxPrepareImageTypeSupport.h"
- Removed `elxPrepareImageTypeSupportMacro()`
- Adjusted `#include` directives in "elxSupportedImageTypes.h.in" accordingly

Aims to ease further maintenance.
