+++ Core/Install/elxMacro.h
@@ -165,70 +165,6 @@
  */
 #define elxInstallComponentFunctionCallMacro(_classname) ret |= _classname##InstallComponent(_cdb)
 
-/**
- * elxPrepareImageTypeSupportMacro
- * To be called once, before the elxSupportImageTypeMacro()
- *
- * IMPORTANT: the macro must be invoked in namespace elastix!
- */
-
-#define elxPrepareImageTypeSupportMacro()                                                                              \
-  template <::elx::ComponentDatabase::IndexType VIndex> /**unsigned int*/                                              \
-  class ElastixTypedef                                                                                                 \
-  {                                                                                                                    \
-  public:                                                                                                              \
-    /** In the specialisations of this template class */                                                               \
-    /** this typedef will make sense */                                                                                \
-    typedef ::itk::Object         ElastixType;                                                                         \
-    constexpr static const char * FixedPixelTypeString{ "" };                                                          \
-    constexpr static const char * MovingPixelTypeString{ "" };                                                         \
-    constexpr static unsigned int FixedDimension{ 0 };                                                                 \
-    constexpr static unsigned int MovingDimension{ 0 };                                                                \
-    /** In the specialisations of this template class, this value will be 'true' */                                    \
-    constexpr static bool IsDefined{ false };                                                                          \
-  }
-
-/**
- * Macro for installing support for new ImageTypes.
- * Used in elxSupportedImageTypes.cxx .
- *
- * Example of usage:
- *
- * namespace elastix {
- * elxSupportedImageTypeMacro(unsigned short, 2, float, 3, 1);
- * elxSupportedImageTypeMacro(unsigned short, 3, float, 3, 2);
- * etc.
- * } //end namespace elastix
- *
- * The first line adds support for the following combination of ImageTypes:
- * fixedImage: 2D unsigned short
- * movingImage 3D float
- * The Index (last argument) of this combination of ImageTypes is 1.
- *
- * The Index should not be 0. This value is reserved for errormessages.
- * Besides, duplicate indices are not allowed.
- *
- * IMPORTANT: the macro must be invoked in namespace elastix!
- *
- * Details: the macro adds a class template specialization for the class
- * ElastixTypedef<VIndex>.
- *
- */
-
-#define elxSupportedImageTypeMacro(_fPixelType, _fDim, _mPixelType, _mDim, _VIndex)                                    \
-  template <>                                                                                                          \
-  class ElastixTypedef<_VIndex>                                                                                        \
-  {                                                                                                                    \
-  public:                                                                                                              \
-    typedef ::itk::Image<_fPixelType, _fDim>                        FixedImageType;                                    \
-    typedef ::itk::Image<_mPixelType, _mDim>                        MovingImageType;                                   \
-    typedef ::elx::ElastixTemplate<FixedImageType, MovingImageType> ElastixType;                                       \
-    constexpr static const char *                                   FixedPixelTypeString{ #_fPixelType };              \
-    constexpr static const char *                                   MovingPixelTypeString{ #_mPixelType };             \
-    constexpr static unsigned int                                   FixedDimension{ _fDim };                           \
-    constexpr static unsigned int                                   MovingDimension{ _mDim };                          \
-    constexpr static bool                                           IsDefined{ true };                                 \
-  }
 
 /**
  * elxClassNameMacro(_name)
